refactor(translations): reorganized translation directory structure
