+++ package.json
@@ -53,7 +53,6 @@
         "escape-html": "^1.0.3",
         "fs-extra": "^8.1.0",
         "glob-promise": "^3.4.0",
-        "i18next": "^19.1.0",
         "jsonpath": "^1.0.2",
         "merge": "^1.2.1",
         "mkdirp": "^1.0.0",
@@ -81,6 +80,8 @@
         "body-parser": "1.19.0",
         "express": "4.17.1",
         "express-fileupload": "1.1.6",
+        "i18next": "^19.1.0",
+        "i18next-node-fs-backend": "^2.1.3",
         "jest": "25.1.0",
         "mockdate": "2.0.5",
         "prettier": "1.19.1",
