--- src/translations/en.player.json
@@ -1,30 +1,0 @@
-{
-    "fullscreen": "Fullscreen",
-    "disableFullscreen": "Disable fullscreen",
-    "download": "Download",
-    "copyrights": "Rights of use",
-    "embed": "Embed",
-    "size": "Size",
-    "showAdvanced": "Show advanced",
-    "hideAdvanced": "Hide advanced",
-    "advancedHelp": "Include this script on your website if you want dynamic sizing of the embedded content:",
-    "copyrightInformation": "Rights of use",
-    "close": "Close",
-    "title": "Title",
-    "author": "Author",
-    "year": "Year",
-    "source": "Source",
-    "license": "License",
-    "thumbnail": "Thumbnail",
-    "noCopyrights": "No copyright information available for this content.",
-    "downloadDescription": "Download this content as a H5P file.",
-    "copyrightsDescription": "View copyright information for this content.",
-    "embedDescription": "View the embed code for this content.",
-    "h5pDescription": "Visit H5P.org to check out more cool content.",
-    "contentChanged": "This content has changed since you last used it.",
-    "startingOver": "You'll be starting over.",
-    "by": "by",
-    "showMore": "Show more",
-    "showLess": "Show less",
-    "subLevel": "Sublevel"
-}
